+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.6.1] - TBD
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
 - [formula-android] Adding main thread check before notifying fragments. 
+- **Breaking**: Crash when duplicate fragment contract is registered.
 
 ## [0.6.0] - July 27, 2020
 - **Breaking**: Changing from RxJava 2.x to RxJava 3.x